Change to service interfaces for tests
