--- kotlin/micronaut/src/main/resources/invalidrow.csv
@@ -1,3 +1,0 @@
-isbn,title,publication_date,genre,author_first,author_last
-ABCEDEFG,Algorithms of Oppression,1605380405,nonfiction,Safiya Umoja,Noble
-123ABC456GEF,Head First Kotlin,1605380405,science,Dawn,Griffiths
